renamed the folder to match name
