+++ src/components/shared/ExternalLinkWithModal/index.tsx
